+++ src/django_components/util/testing.py
@@ -17,6 +17,7 @@ from django_components.component import ALL_COMPONENTS, Component, component_nod
 from django_components.component_media import ComponentMedia
 from django_components.component_registry import ALL_REGISTRIES, ComponentRegistry
 from django_components.extension import extensions
+from django_components.perfutil.provide import provide_cache
 from django_components.template import _reset_component_template_file_cache, loading_components
 
 # NOTE: `ReferenceType` is NOT a generic pre-3.9
@@ -471,6 +472,9 @@ def _clear_djc_global_state(
 
     if component_media_cache:
         component_media_cache.clear()
+
+    if provide_cache:
+        provide_cache.clear()
 
     # Remove cached Node subclasses
     component_node_subclasses_by_name.clear()